requirements.txt
requests>=2.31.0
openpyxl>=3.1.0
gspread>=6.0.0
google-auth>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pytest>=8.0.0
